Move V3 assessment entities to bare module name
